+++ src/TextShow/GHC/Stats.hs
@@ -1,7 +1,10 @@
 {-# LANGUAGE CPP             #-}
+
+#if !(MIN_VERSION_base(4,11,0))
 {-# LANGUAGE TemplateHaskell #-}
 {-# OPTIONS_GHC -fno-warn-deprecations #-}
 {-# OPTIONS_GHC -fno-warn-orphans      #-}
+#endif
 
 {-|
 Module:      TextShow.GHC.Stats
@@ -17,6 +20,7 @@ Portability: GHC
 -}
 module TextShow.GHC.Stats () where
 
+#if !(MIN_VERSION_base(4,11,0))
 import GHC.Stats (GCStats)
 
 import TextShow.Data.Integral ()
@@ -25,3 +29,4 @@ import TextShow.TH.Internal (deriveTextShow)
 
 -- /Since: 2/
 $(deriveTextShow ''GCStats)
+#endif