+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### next [????.??.??]
+* Don't define a `TextShow GCStats` instance on GHC 8.4 or later, as `GCStats`
+  has been removed from `base`. This fixes the build on GHC 8.4.
+
 ## 3.7 [2017.12.07]
 * Use `EmptyCase` on GHC 7.8 or later for `TextShow(1)` instances for empty data types that are derived using `TextShow.Generic`.
 * Derived `TextShow(1)(2)` instances (using `TextShow.TH`) will now force their argument instead of simply `error`ing.